requirements.txt
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
dnspython==2.5.0
Flask==3.0.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonify==0.5
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
tzdata==2023.4
Werkzeug==3.0.1
